requirements.txt
Django==4.0.5
Pillow==9.3.0
psycopg2-binary==2.9.5
django-cors-headers==4.3.0
django-rest-authtoken==2.1.4
djangorestframework==3.14.0
gunicorn==21.2.0